==> Notebooks/5_moh_source code implementation.ipynb <==
# 1- Imports (Source Code + Libraries)

import pickle
import numpy as np
import pandas as pd

import os

# adding source code parent directory
import sys
sys.path.append('..\\Src')

### 1.1 Data Preprocessing Modules

from preprocessing.data_prep import read_data
from preprocessing.data_clean import clean_data2, clean_text
from preprocessing.feature_extract import create_vectorizer

### 1.2 Search Engine Module

from search.search_engine import Search_Engine

### 1.3 utils

from utils.modify_stop_words import modify_sw

# 2- Preparation for SE

# variables set
data_path = "../Data/husna.json"
cols = ['text_clean', 'summary_clean', 'title_clean', 'tags']
cols_weights = {'text_clean': 0.5, 'summary_clean': 0.15, 'title_clean': 0.2, 'tags': 0.1}

# prepare stop words (use when list.csv is not available)
# modify_sw(path='../Data/stop_words/stop_words_list.txt')

# import -> clean data
docs_df = read_data(data_path)
docs_df_cleaned = clean_data2(docs_df)    

docs_df.shape

vectorizers = {}

for col in cols[:-1]:
    # create vectorizer for each column
    vectorizers[col] = create_vectorizer(docs_df_cleaned, from_feature=col)
    
    # saving vectorizers for SE preprocessing
    vectorizer_to_save = vectorizers[col]
    if vectorizer_to_save != None:
        pickle.dump(vectorizer_to_save, open(f"../Results/saved_vectorizers/{col}_tfidf.pickle", "wb"))
        
# saving cleaned data
docs_df_cleaned_SE = docs_df_cleaned[['_id', 'url', 'title', 'text', 'published_at', 'article_type', 'tags', 'text_clean', 'summary_clean', 
                                              'title_clean']]
docs_df_cleaned_SE.to_csv('../Data/processed/SE_data1.csv', index=False, encoding='utf-8-sig')

# 3- using SE

!python -m pip install elasticsearch

SE_df = pd.read_csv('../Data/processed/SE_data1.csv', keep_default_na=False, encoding='utf-8-sig')

SE_df

SE_ob = Search_Engine(cols, cols_weights, df=SE_df)

SE_ob.get_saved_vectorizers(rel_dir='../Results/saved_vectorizers')

results = SE_ob.overall_search('مولد النبي')

np.array(top10_results, dtype="int32")[:, 0]

ids = np.array(top10_results, dtype="int32")[:, 0]
urls_temp = list(SE_df.loc[ids, :]['url'])
title_temp = list(SE_df.loc[ids, :]['title'])
text_temp = list(SE_df.loc[ids, :]['text'])

q = 'التعليم'
clean_text(q)

top10_results = sorted(results[0], key=lambda x: x[1], reverse=True)[:10]
SE_time = results[1]
SE_score = results[2]
ids = np.array(top10_results, dtype="int32")[:, 0]
urls = list(SE_df.loc[ids, :]['url'])
titles = list(SE_df.loc[ids, :]['title'])
texts = list(SE_df.loc[ids, :]['text'])

print(f'-> Top 10 Results:')
print('-' * 100, end='\n\n')
for i, (t10r_id, t10r_score) in enumerate(top10_results):
    print('Document ID {:d} with a score of {:.3f}'.format(int(t10r_id), t10r_score))
    print(f"link: {urls[i]}", end='\n\n')
    print(f"title: {titles[i]}")
    print(f"description: {texts[i][:100]}")
    print('-' * 50, end='\n\n')
print()
print('Search Time Duration: {:.2f} Seconds'.format(SE_time * 10 **-3))
print('Search (uniqueness/frequency) fraction: {:.3f}'.format(SE_score))
