==> Notebooks/6_moh_SE_elastic_search.ipynb <==
## 1- Library and Data Imports

import numpy as np
import pandas as pd
import time

# for text cleaning and preprocessing
import re
from nltk.corpus import stopwords
import string 
from sklearn.feature_extraction.text import TfidfVectorizer

import sys
sys.path.append('../Src')
from elasticsearch import Elasticsearch
from elasticsearch.client import IndicesClient
from preprocessing.data_clean import clean_text

SE_df = pd.read_csv('../Data/processed/SE_data1.csv', keep_default_na=False, encoding='utf-8-sig')

SE_df

docs_to_store = SE_df[['url', 'title', 'text', 'tags']]
docs_to_store = docs_to_store.to_dict('records')

## 2- Connection 

import logging
def connect_elasticsearch():
    _es = None
    ELASTIC_PASSWORD = "oC+QPrfW=4B8GpDC_1hr"
    _es = Elasticsearch("https://localhost:9200", 
                   ca_certs='../Certs/http_ca.crt',
                   basic_auth=("elastic", ELASTIC_PASSWORD))
    if _es.ping():
        print('Yay Connect')
    else:
        print('Awww it could not connect!')
    return _es

es = connect_elasticsearch()

# from datetime import datetime
# from elasticsearch import Elasticsearch


# ELASTIC_PASSWORD = "oC+QPrfW=4B8GpDC_1hr"
# es = Elasticsearch("https://localhost:9200", 
# #                    client_cert='a42181244c443b87e84571fb7d9b630ef550537de80636eac298e7ec8abbde33',
# #                    verify_certs=True,
#                    ca_certs='../Certs/http_ca.crt',
#                    basic_auth=("elastic", ELASTIC_PASSWORD))

# es.ping()

def create_index(es_object, index_name='SE_SHAI'):
    created = False
    # index settings
    settings = {
        "settings": {
            "number_of_shards": 1,
            "number_of_replicas": 0
        },
        "mappings": {
            "members": {
                "dynamic": "strict",
                "properties": {
                    "url": {
                        "type": "text"
                    },
                    "title": {
                        "type": "text"
                    },
                    "text": {
                        "type": "text"
                    },
                    "tags": {
                        "type": "text"
                    },
                }
            }
        }
    }
    
    try:
        if not es_object.indices.exists(index=index_name):
            # Ignore 400 means to ignore "Index Already Exist" error.
            es_object.indices.create(index=index_name, ignore=400, body=settings)
            print('Created Index')
        else:
            print('index exists')
        created = True
    except Exception as ex:
        print(str(ex))
    finally:
        return created

print(f"index created: {create_index(es, 'se_shai')}")

def store_record(elastic_object, index_name, record, record_id):
    try:
        outcome = elastic_object.index(index=index_name, id=record_id, document=record)
    except Exception as ex:
        print('Error in indexing data')
        print(str(ex))

for i, doc in enumerate(docs_to_store):
    store_record(es, 'se_shai', doc, i)

def get_records(elastic_object, index_name, record_id):
    try:
        resp = elastic_object.get(index=index_name, id=record_id)
        return resp
    except Exception as ex:
        print('Error in indexing data')
        print(str(ex))

resp_ret = get_records(es, 'se_shai', 1)
print(resp_ret)

# refereshing indices

# es.indices.refresh(index="se_shai")

def search(es_object, index_name, search):
    res = es_object.search(index=index_name, query=search)

- for search, BM25 similarity is applied for scoring matched documents.
- scoring / ranking model: **BM25**, which is a TF/IDF based similarity that has built-in tf normalization.
    - other similarity ranking approaches: DFR similarity, DFI similarity, IB similarity, LM Dirichlet similarity, LM Jelinek Mercer similarity, Scripted similarity (Custom).  

q = 'مولد النبي'

resp = es.search(index="se_shai", query=
                 {"multi_match": 
                  { 
                      "query": q, 
                      "fields": [ "title", "text", "tags"] 
                  }
                 }
                )

# list of documents with scores
resp['hits']['hits']
# print("Got %d Hits:" % resp['hits']['total']['value'])
# print("Got %d Hits:" % resp['hits']['total']['value'])
# for hit in resp['hits']['hits']:
#     print("%(timestamp)s %(author)s: %(text)s" % hit["_source"])

i_es = IndicesClient(es)

# get index
i_es.get(index='se_shai*')

# delete index 
i_es.delete(index='se_shai')


es.close()

## Building Search Functionality

from elasticsearch import Elasticsearch
from elasticsearch.client import IndicesClient

def connect_elasticsearch():
    _es = None
    ELASTIC_PASSWORD = "oC+QPrfW=4B8GpDC_1hr"
    _es = Elasticsearch("https://localhost:9200", 
                   ca_certs='../Certs/http_ca.crt',
                   basic_auth=("elastic", ELASTIC_PASSWORD))
    if _es.ping():
        print('Yay Connect')
    else:
        print('Awww it could not connect!')
    return _es

es = connect_elasticsearch()

q = 'مقال'
q_cleaned = clean_text(q)
q_cleaned

query = {"multi_match": 
                  { 
                      "query":  q+'*', 
                      "fields": [ "title", "text", "tags"] 
                  }
}

resp = es.search(index="se_shai", query=query)

urls = []
titles = []
texts = []
tags = []

# list of documents with scores
for hit in resp['hits']['hits']:
    urls.append(hit['_source']['url'])
    titles.append(hit['_source']['title'])
    texts.append(hit['_source']['text'])
    tags.append(hit['_source']['tags'])
